check if message id deletable

+++ src/commands/help.js
@@ -12,6 +12,8 @@ module.exports = {
         });
 
         msg.author.send(embed);
-        msg.original.delete();
+        if (msg.original.deletable) {
+            msg.original.delete();
+        }
     }
 };
